+++ tests/test_create_observation/test_create_software_version_observation/mock_info.py
@@ -1,0 +1,24 @@
+"""MockInfo"""
+
+from typing import List
+
+
+class MockBackgroundClassList:
+    """Mock Backround Class List"""
+
+    tasks: List
+
+    def add_task(self, task):
+        self.tasks.append(task)
+
+    def __init__(self):
+        self.tasks = []
+
+
+class MockInfo:
+    """Mock StrawberryInfo"""
+
+    context: dict
+
+    def __init__(self, context: dict):
+        self.context = context
